WIP use just GOBIN not GOPATH

+++ test/e2e/e2e-golang.sh
@@ -7,12 +7,11 @@ source $(dirname $0)/e2e-paths.env
 GO_VERSION=1.11.4
 
 echo ">>> Installing go ${GO_VERSION} to $GOBASE/go"
-curl -O https://storage.googleapis.com/golang/go${GO_VERSION}.linux-amd64.tar.gz
+if ! [ -f "go${GO_VERSION}.linux-amd64.tar.gz" ]; then
+    curl -O https://storage.googleapis.com/golang/go${GO_VERSION}.linux-amd64.tar.gz
+fi
 tar -xf go1.11.4.linux-amd64.tar.gz
 rm -rf $GOBASE/go
 mv go $GOBASE/
-
-mkdir -p $GOPATH/bin
-mkdir -p $GOPATH/src
 
 go version
